requirements.txt
fastapi>=0.109.2
uvicorn>=0.25.0
numpy>=1.26.3
python-dotenv>=1.0.1
pydantic>=2.5.3
loguru>=0.7.2